+++ CHANGELOG.md
@@ -8,6 +8,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ## [Unreleased]
 
 
+### Added
+- Print hashes.
+
+
 ## [0.0.2] - 2018-07-10
 ### Changed
 - Move to new zenbuild.